pyproject.toml
[project]
name = "dash-py-mercado-financeiro"
version = "0.1.0"
description = ""
authors = [
    {name = "leovnoliveira",email = "consultoria1.ceo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "selenium (>=4.29.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "lxml (>=5.3.1,<6.0.0)",
    "metatrader5 (>=5.0.4874,<6.0.0)",
    "pyarrow (>=19.0.1,<20.0.0)",
    "fastparquet (>=2024.11.0,<2025.0.0)",
    "python-bcb (>=0.3.2,<0.4.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "dash (>=2.18.2,<3.0.0)",
    "dash-bootstrap-components (>=1.7.1,<2.0.0)",
    "components (>=0.0.1a0,<0.0.2)"
]

[tool.poetry]
packages = [{include = "dash_py_mercado_financeiro", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
